Move history reducer into reducer.ts
